+++ packages/core/docs/changelog/5363.js
@@ -1,8 +1,0 @@
-module.exports = {
-  description: 'added useUiNotification composable',
-  link: 'https://github.com/vuestorefront/vue-storefront/issues/5363',
-  isBreaking: false,
-  breakingChanges: [],
-  author: 'Łukasz Jędrasik',
-  linkToGitHubAccount: 'https://github.com/lukaszjedrasik'
-};